backend/flask-src/requirements.txt
flask
numpy
pymongo
python-dotenv
pandas
scipy
scikit-learn
matplotlib
statsmodels
